added week 09 example from class

+++ Week09/09MoCapMagic/src/Trail.h
@@ -5,15 +5,13 @@ class Trail
 {
 public:
 	Trail();
-	Trail(float x, float y, ofColor c, float r);
-
+	Trail(float x, float y, ofColor c, int r);
 	bool update();
 	void display();
 
 	ofVec2f position;
 	ofColor color;
 	int transparency = 255;
-	float radius;
-
+	int radius;
 };
